requirements.txt
accelerate==1.13.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.13.0
attrs==26.1.0
audioread==3.1.0
av==17.0.0
bitsandbytes==0.49.2
boto3==1.42.74
botocore==1.42.74
brotli==1.2.0
cached-path==1.8.10
causal-conv1d @ https://github.com/Dao-AILab/causal-conv1d/releases/download/v1.6.1.post4/causal_conv1d-1.6.1+cu12torch2.8cxx11abiTRUE-cp311-cp311-linux_x86_64.whl
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
contourpy==1.3.3
crcmod==1.7
cryptography==46.0.5
ctranslate2==4.7.1
cycler==0.12.1
cython==0.29.37
datasets==4.8.4
decorator==5.2.1
dill==0.4.1
editdistance==0.8.1
einops==0.8.2
einx==0.4.2
ema-pytorch==0.7.9
encodec==0.1.1
-e file:///home/anekazek/anything-notsupported-inwindows/kcv-tts
fastapi==0.135.2
faster-whisper==1.2.1
ffmpy==1.0.0
filelock==3.25.2
fire==0.7.1
flatbuffers==25.12.19
fonttools==4.62.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2026.2.0
funasr==1.3.1
gitdb==4.0.12
gitpython==3.1.46
google-api-core==2.30.0
google-auth==2.49.1
google-cloud-core==2.5.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
gradio==6.9.0
gradio-client==2.3.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.7.2
hydra-core==1.3.2
idna==3.11
jaconv==0.5.0
jamo==0.4.1
jieba==0.42.1
jinja2==3.1.6
jiwer==4.0.0
jmespath==0.10.0
joblib==1.5.3
kaldiio==2.18.1
kiwisolver==1.5.0
lazy-loader==0.5
librosa==0.11.0
llvmlite==0.46.0
loguru==0.7.3
mamba-ssm @ https://github.com/state-spaces/mamba/releases/download/v2.3.1/mamba_ssm-2.3.1+cu12torch2.8cxx11abiTRUE-cp311-cp311-linux_x86_64.whl
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
modelscope==1.35.1
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multiprocess==0.70.19
networkx==3.6.1
ninja==1.13.0
numba==0.64.0
numpy==2.4.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
onnxruntime==1.24.4
orjson==3.11.7
oss2==2.19.1
packaging==26.0
pandas==3.0.1
pillow==12.1.1
pip==26.0.1
platformdirs==4.9.4
pooch==1.9.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.6
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-core==2.41.5
pydub==0.25.1
pygments==2.19.2
pynndescent==0.6.0
pyparsing==3.3.2
pypinyin==0.55.0
python-dateutil==2.9.0.post0
python-fire==0.1.0
python-multipart==0.0.22
pytorch-wpe==0.0.1
pytz==2026.1.post1
pyyaml==6.0.3
rapidfuzz==3.14.3
regex==2026.2.28
requests==2.32.5
rich==13.9.4
rjieba==0.2.0
s3transfer==0.16.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
semantic-version==2.10.0
sentencepiece==0.2.1
sentry-sdk==2.56.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
soundfile==0.13.1
soxr==1.0.0
starlette==0.52.1
sympy==1.14.0
tensorboardx==2.6.4
termcolor==3.3.0
threadpoolctl==3.6.0
tokenizers==0.22.2
tomli==2.4.0
tomlkit==0.13.3
torch==2.8.0+cu128
torch-complex==0.4.4
torchaudio==2.8.0+cu128
torchcodec==0.11.0
torchdiffeq==0.2.5
tqdm==4.67.3
transformers==5.3.0
transformers-stream-generator==0.0.5
triton==3.4.0
typer==0.24.1
typing-extensions==4.15.0
typing-inspection==0.4.2
umap-learn==0.5.11
unidecode==1.4.0
urllib3==2.6.3
uvicorn==0.42.0
vocos==0.1.0
wandb==0.25.1
wheel==0.46.3
x-transformers==2.17.7
xxhash==3.6.0
yarl==1.23.0
zhconv==1.4.3
zhon==2.1.1
